% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_price.R
\name{get_all_stock_prices}
\alias{get_all_stock_prices}
\title{Read all stock prices from Azure SQL in a single database read.}
\usage{
get_all_stock_prices(con)
}
\arguments{
\item{con}{database connection}
}
\description{
Read all stock prices from Azure SQL in a single database read.
}
