% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftse100_components.R
\name{get_ftse100_components}
\alias{get_ftse100_components}
\title{Read FTSE 100 components and tickers into data frame.}
\usage{
get_ftse100_components(con)
}
\arguments{
\item{con}{database connection}
}
\value{
tibble
}
\description{
Read FTSE 100 components and tickers into data frame.
}
