% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_price.R
\name{get_stock_price}
\alias{get_stock_price}
\title{Read stock prices from Azure SQL.}
\usage{
get_stock_price(con, ticker)
}
\arguments{
\item{con}{database connection}

\item{ticker}{ticker}
}
\value{
tibble with ticker price data
}
\description{
Read stock prices from Azure SQL.
}
