% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftse100_components.R
\name{scrape_ftse100_components}
\alias{scrape_ftse100_components}
\title{Get list of FTSE 100 index components from wikipedia.}
\usage{
scrape_ftse100_components()
}
\value{
tibble
}
\description{
Scrapes "https://en.wikipedia.org/wiki/FTSE_100_Index" for list of index components
}
