% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_price.R
\name{scrape_stock_price}
\alias{scrape_stock_price}
\title{Get stock prices from yahoo finance.}
\usage{
scrape_stock_price(ticker, wait = 0.1)
}
\arguments{
\item{ticker}{ticker}

\item{wait}{time in seconds to wait between yahoo reads}
}
\value{
tibble with ticker price data
}
\description{
Use \code{quantmod} to read ticker data from yahoo finance.
}
