% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftse100_components.R
\name{update_ftse_components}
\alias{update_ftse_components}
\title{Scrape FTSE component and wirte to database.}
\usage{
update_ftse_components(con, tickers)
}
\arguments{
\item{con}{database connection}

\item{tickers}{Stock tickers}
}
\description{
Scrape FTSE component and wirte to database.
}
