% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_price.R
\name{update_stock_price}
\alias{update_stock_price}
\title{Read stock prices from Yahoo Finance and update Azure SQL.}
\usage{
update_stock_price(con, ticker, wait = 0.1)
}
\arguments{
\item{con}{database connection}

\item{ticker}{ticker}

\item{wait}{time in seconds to wait between yahoo reads}
}
\value{
tibble with ticker price data
}
\description{
Read stock prices from Yahoo Finance and update Azure SQL.
}
