% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_connect.R
\name{connect_to_azure_sql}
\alias{connect_to_azure_sql}
\title{Connect to Azure SQL database.}
\usage{
connect_to_azure_sql(db, id, pw, port = 1433)
}
\arguments{
\item{db}{database server name, e.g. \code{tcp:adv-stock-price.database.windows.net}}

\item{id}{user name}

\item{pw}{password}

\item{port}{port number}
}
\value{
database connection string
}
\description{
Connect to Azure SQL database.
}
